(#22) ✨ style: 추천 구역 관련 데이터를 Enum으로 관리
