+++ src/constants/ZoneData.ts
@@ -1,0 +1,36 @@
+// Stadium 관리
+export enum StadiumType {
+    NONE = '',
+    JAMSIL = '잠실종합운동장',
+    KT = '수원KT위즈파크'
+}
+
+// 잠실 Zone 관리
+
+// KT Zone 관리
+
+// Seat 관리
+export enum SeatType {
+    NONE = '',
+    SEAT1 = '1루석',
+    SEAT3 = '3루석'
+}
+
+// keyword 관리
+export enum Keyword {
+    NONE = '',
+    PARTNER1 = '나 혼자',
+    PARTNER2 = '같은 팀 팬과',
+    PARTNER3 = '다른 팀 팬과',
+    WISH1 = '열정적인 응원',
+    WISH2 = '경기장 한눈에 보기',
+    WISH3 = '편리한 화장실 이용',
+    WISH4 = '음식 먹기 편한',
+    WISH5 = '빠른 퇴장 가능',
+    WISH6 = '선수들 가까이',
+    WISH7 = '삼겹살 구워먹기',
+    NOWISH1 = '햇빛 싫어요',
+    NOWISH2 = '큰 소리 싫어요',
+    NOWISH3 = '높은 곳 싫어요',
+    NOWISH4 = '비 맞기 싫어요'
+}
